{{cookiecutter.repo_name}}/requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==0.9.4
Sphinx
